move filtering with io functions into `filtering` away from cli
